requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Database
supabase>=2.0.0

# API server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Payment processing
stripe>=7.0.0

# Utilities
urllib3>=1.26.0
lxml>=4.9.0
httpx>=0.24.0
pydantic>=2.0.0

# Payment processing
stripe>=5.0.0
